feat: check pair balance
Disable pair when opener < closer, close when opener > closer, and
other actions when unbalanced. Pairs backslash-escaped and in quoted
strings will be ignored when checking pair balance. Types of quotes
are stored in a separate file for some filetypes.

Use a scanner instead of lua pattern or vim regex to count number of
openers and closers because a scanner seems to be more performative
according to my rough benchmarking.